requirements.txt
Faker==40.15.0
Flask==3.1.3
mysql_connector_repackaged==0.3.1
networkx==3.6.1
pandas==3.0.2
pulp==3.3.0
